dashboard/requirements.txt
streamlit==1.35.0
pandas>=2.0.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
requests==2.32.3
altair>=5.0.0
